// src/types/file/SongRowSave.ts
export default interface SongRowSave {
  songID: number;
  lighterRating: number;
  playCount: number;

  drumsTopScore: number;
  drumsTopScoreDifficulty: number;
  drumsStarsEasy: number;
  drumsPercentEasy: number;
  drumsStarsMedium: number;
  drumsPercentMedium: number;
  drumsStarsHard: number;
  drumsPercentHard: number;
  drumsStarsExpert: number;
  drumsPercentExpert: number;

  bassTopScore: number;
  bassTopScoreDifficulty: number;
  bassStarsEasy: number;
  bassPercentEasy: number;
  bassStarsMedium: number;
  bassPercentMedium: number;
  bassStarsHard: number;
  bassPercentHard: number;
  bassStarsExpert: number;
  bassPercentExpert: number;

  guitarTopScore: number;
  guitarTopScoreDifficulty: number;
  guitarStarsEasy: number;
  guitarPercentEasy: number;
  guitarStarsMedium: number;
  guitarPercentMedium: number;
  guitarStarsHard: number;
  guitarPercentHard: number;
  guitarStarsExpert: number;
  guitarPercentExpert: number;

  vocalsTopScore: number;
  vocalsTopScoreDifficulty: number;
  vocalsStarsEasy: number;
  vocalsPercentEasy: number;
  vocalsStarsMedium: number;
  vocalsPercentMedium: number;
  vocalsStarsHard: number;
  vocalsPercentHard: number;
  vocalsStarsExpert: number;
  vocalsPercentExpert: number;

  harmoniesTopScore: number;
  harmoniesTopScoreDifficulty: number;
  harmoniesStarsEasy: number;
  harmoniesPercentEasy: number;
  harmoniesStarsMedium: number;
  harmoniesPercentMedium: number;
  harmoniesStarsHard: number;
  harmoniesPercentHard: number;
  harmoniesStarsExpert: number;
  harmoniesPercentExpert: number;

  keysTopScore: number;
  keysTopScoreDifficulty: number;
  keysStarsEasy: number;
  keysPercentEasy: number;
  keysStarsMedium: number;
  keysPercentMedium: number;
  keysStarsHard: number;
  keysPercentHard: number;
  keysStarsExpert: number;
  keysPercentExpert: number;

  proDrumsTopScore: number;
  proDrumsTopScoreDifficulty: number;
  proDrumsStarsEasy: number;
  proDrumsPercentEasy: number;
  proDrumsStarsMedium: number;
  proDrumsPercentMedium: number;
  proDrumsStarsHard: number;
  proDrumsPercentHard: number;
  proDrumsStarsExpert: number;
  proDrumsPercentExpert: number;

  proGuitarTopScore: number;
  proGuitarTopScoreDifficulty: number;
  proGuitarStarsEasy: number;
  proGuitarPercentEasy: number;
  proGuitarStarsMedium: number;
  proGuitarPercentMedium: number;
  proGuitarStarsHard: number;
  proGuitarPercentHard: number;
  proGuitarStarsExpert: number;
  proGuitarPercentExpert: number;

  proBassTopScore: number;
  proBassTopScoreDifficulty: number;
  proBassStarsEasy: number;
  proBassPercentEasy: number;
  proBassStarsMedium: number;
  proBassPercentMedium: number;
  proBassStarsHard: number;
  proBassPercentHard: number;
  proBassStarsExpert: number;
  proBassPercentExpert: number;

  proKeysTopScore: number;
  proKeysTopScoreDifficulty: number;
  proKeysStarsEasy: number;
  proKeysPercentEasy: number;
  proKeysStarsMedium: number;
  proKeysPercentMedium: number;
  proKeysStarsHard: number;
  proKeysPercentHard: number;
  proKeysStarsExpert: number;
  proKeysPercentExpert: number;

  bandTopScore: number;
  bandTopScoreDifficulty: number;
  bandStarsEasy: number;
  bandPercentEasy: number;
  bandStarsMedium: number;
  bandPercentMedium: number;
  bandStarsHard: number;
  bandPercentHard: number;
  bandStarsExpert: number;
  bandPercentExpert: number;
}